backend/pyproject.toml
[project]
name = "multi-phase-todo-backend"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "asyncpg==0.29.0",
    "alembic==1.12.1",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "httpx==0.25.2",
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "ruff==0.1.6",
    "black==23.10.1",
    "mypy==1.7.1",
    "slowapi>=0.1.9",
    "click>=8.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "ruff==0.1.6",
    "black==23.10.1",
    "mypy==1.7.1"
]

# Tell uv which package directory to install
[tool.uv.build]
packages = ["multi_phase_todo_backend"]  # <-- your Python package folder (with __init__.py)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--tb=short",
    "-v"
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | node_modules
)/
'''

[tool.ruff]
select = [
    "E",
    "W",
    "F",
    "I",
    "C",
    "B"
]
line-length = 88
target-version = "py312"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
